M to M relationship on .net 6 #3
