+++ src/Sample.App/Jobs/JobBase.cs
@@ -5,23 +5,22 @@ using Microsoft.Extensions.Logging;
 
 using Sample.Data;
 
-namespace Sample.App.Jobs
+namespace Sample.App.Jobs;
+
+public abstract class JobBase
 {
-    public abstract class JobBase
+    public JobBase(AppDbContext context, ILogger logger)
     {
-        public JobBase(AppDbContext context, ILogger logger)
-        {
-            this.context = context ?? throw new ArgumentNullException(nameof(context));
-            this.logger = logger ?? throw new ArgumentNullException(nameof(logger));
-        }
-
-        protected AppDbContext Context => context;
-
-        protected ILogger Logger => logger;
-
-        public abstract Task ExecuteAsync();
-
-        private readonly AppDbContext context;
-        private readonly ILogger logger;
+        this.context = context ?? throw new ArgumentNullException(nameof(context));
+        this.logger = logger ?? throw new ArgumentNullException(nameof(logger));
     }
+
+    protected AppDbContext Context => context;
+
+    protected ILogger Logger => logger;
+
+    public abstract Task ExecuteAsync();
+
+    private readonly AppDbContext context;
+    private readonly ILogger logger;
 }
