+++ src/Sample.App/Jobs/QuerySampleDataJob.cs
@@ -7,35 +7,34 @@ using Microsoft.Extensions.Logging;
 using Sample.App.Features;
 using Sample.Data;
 
-namespace Sample.App.Jobs
+namespace Sample.App.Jobs;
+
+public class QuerySampleDataJob : JobBase
 {
-    public class QuerySampleDataJob : JobBase
+    public QuerySampleDataJob(ObjectViewer viewer, AppDbContext context, ILogger<QuerySampleDataJob> logger) : base(context, logger)
     {
-        public QuerySampleDataJob(ObjectViewer viewer, AppDbContext context, ILogger<QuerySampleDataJob> logger) : base(context, logger)
-        {
-            this.viewer = viewer;
-        }
-
-        public override async Task ExecuteAsync()
-        {
-            var students = await Context.Students
-                .Include(x => x.Enrollments)
-                .Select(x => new
-                {
-                    Name = x.Name,
-                    Courses = x.Enrollments.Select(x => new
-                    {
-                        Title = x.Course.Title,
-                    }),
-                }).ToListAsync();
-
-            //Logger.LogInformation("Student: {@students}", students);
-
-            //System.Console.WriteLine(viewer.Print(students));
-
-            Logger.LogInformation("Student: {students}", viewer.PringJson(students));
-        }
-
-        private readonly ObjectViewer viewer;
+        this.viewer = viewer;
     }
+
+    public override async Task ExecuteAsync()
+    {
+        var students = await Context.Students
+            .Include(x => x.Enrollments)
+            .Select(x => new
+            {
+                Name = x.Name,
+                Courses = x.Enrollments.Select(x => new
+                {
+                    Title = x.Course.Title,
+                }),
+            }).ToListAsync();
+
+        //Logger.LogInformation("Student: {@students}", students);
+
+        //System.Console.WriteLine(viewer.Print(students));
+
+        Logger.LogInformation("Student: {students}", viewer.PringJson(students));
+    }
+
+    private readonly ObjectViewer viewer;
 }
