+++ src/Sample.Data/AppDbContext.cs
@@ -4,22 +4,21 @@ using Microsoft.EntityFrameworkCore;
 
 using Sample.Entities;
 
-namespace Sample.Data
+namespace Sample.Data;
+
+public class AppDbContext : DbContext
 {
-    public class AppDbContext : DbContext
+    public AppDbContext(DbContextOptions<AppDbContext> options) : base(options)
     {
-        public AppDbContext(DbContextOptions<AppDbContext> options) : base(options)
-        {
 
-        }
+    }
 
-        public DbSet<Student> Students { get; set; }
+    public DbSet<Student> Students { get; set; }
 
-        public DbSet<Course> Courses { get; set; }
+    public DbSet<Course> Courses { get; set; }
 
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.ApplyConfigurationsFromAssembly(GetType().Assembly);
-        }
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.ApplyConfigurationsFromAssembly(GetType().Assembly);
     }
 }
