--- src/Sample.Data/EntityTypeConfigurations/EnrollmentEntityTypeConfiguration.cs
@@ -1,30 +1,0 @@
-﻿
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-using Sample.Entities;
-
-namespace Sample.Data.EntityTypeConfigurations
-{
-    public class EnrollmentEntityTypeConfiguration : IEntityTypeConfiguration<Enrollment>
-    {
-        public void Configure(EntityTypeBuilder<Enrollment> builder)
-        {
-            builder.HasKey(x => new { x.StudentId, x.CourseId });
-
-            builder.Property(x => x.StudentId)
-                .IsRequired();
-
-            builder.Property(x => x.CourseId)
-                .IsRequired();
-
-            builder.HasOne(x => x.Student)
-                .WithMany(x => x.Enrollments)
-                .HasForeignKey(x => x.StudentId);
-
-            builder.HasOne(x => x.Course)
-                .WithMany(x => x.Enrollments)
-                .HasForeignKey(x => x.CourseId);                
-        }
-    }
-}
