+++ src/Sample.Data/EntityTypeConfigurations/StudentEntityTypeConfiguration.cs
@@ -7,20 +7,19 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 using Sample.Entities;
 
-namespace Sample.Data.EntityTypeConfigurations
+namespace Sample.Data.EntityTypeConfigurations;
+
+public class StudentEntityTypeConfiguration : IEntityTypeConfiguration<Student>
 {
-    public class StudentEntityTypeConfiguration : IEntityTypeConfiguration<Student>
+    public void Configure(EntityTypeBuilder<Student> builder)
     {
-        public void Configure(EntityTypeBuilder<Student> builder)
-        {
-            builder.HasKey(x => x.Id);
+        builder.HasKey(x => x.Id);
 
-            builder.Property(x => x.Id)
-                .IsRequired()
-                .ValueGeneratedOnAdd();
+        builder.Property(x => x.Id)
+            .IsRequired()
+            .ValueGeneratedOnAdd();
 
-            builder.Property(x => x.Name)
-                .IsRequired();
-        }
+        builder.Property(x => x.Name)
+            .IsRequired();
     }
 }
